refactor: 장바구니 상품 기준의 skeleton 반영 - 해당 기준의 컴포넌트 관심사 분리
